added logback and listner'

+++ src/main/java/com/maverick/rmqscheduler/dto/ApiResponseDto.java
@@ -17,5 +17,5 @@ public class ApiResponseDto {
 
     private String uuid;
 
-    private Integer delay;
+    private Long delay;
 }
