+++ src/main/java/com/maverick/rmqscheduler/dto/MessageRequestDto.java
@@ -15,5 +15,5 @@ public class MessageRequestDto {
 
     private String message;
 
-    private Integer delayInMs;
+    private Long delayInMs;
 }
